feat(net): move status updating logic to NetworkManager and NetworkHandle
